Allow "zipcode" in addressbatch()
Currently, the `address()` function expects the keyword argument "zipcode". However, the `addressbatch()` function, if you use the iterable-of-dicts form of argument, expects each dict to have the key "zip". This makes it so users of this library have to use two different schemas, "zip" in one place and "zipcode" in another, depending on if they use `address()` or `addressbatch()`.

This tweaks `addressbatch()` so if you use an iterable-of-dicts, you can use the key "zipcode" instead of "zip". Now in my user code, I only need to have one schema, the one with "zipcode".

This change *doe not* affect `addressbatch()` when you pass in a CSV file. One, I didn't see an easy way to fix this (first thought would be to wrap the file-like-object in a custom file-like-object that modifies the contents on every call to `read()`). Two, I'd imagine the need isn't as serious here, because if you're using a file-like-object then I'd think that you wouldn't even be writing the code that formats that data.